requirements.txt
asgiref==3.7.0
black==23.3.0
boto3==1.26.142
botocore==1.29.142
cffi==1.15.1
click==8.1.3
colorama==0.4.6
coverage==7.2.5
Django==4.2.1
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
jmespath==1.0.1
Markdown==3.4.3
model-mommy==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
pycparser==2.21
Pillow==9.5.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.0
tzdata==2023.3
urllib3==1.26.16
zope.event==4.6
zope.interface==6.0
